Move controllerengine files into new engine subfolder
