+++ src/control/controlobjectscript.h
@@ -1,7 +1,7 @@
 #ifndef CONTROLOBJECTSCRIPT_H
 #define CONTROLOBJECTSCRIPT_H
 
-#include "controllers/controllerengine.h"
+#include "controllers/engine/controllerengine.h"
 #include "controllers/controllerdebug.h"
 #include "control/controlproxy.h"
